Fix precision bug for product price between Kotlin & Postgres

+++ src/main/resources/db/migration/V1__schema.sql
@@ -1,8 +1,9 @@
 DROP TABLE IF EXISTS product;
 
-CREATE TABLE product (
+CREATE TABLE product
+(
     id          BIGSERIAL PRIMARY KEY,
     name        VARCHAR(255) NOT NULL,
-    description TEXT NOT NULL,
-    price       DECIMAL NOT NULL
+    description TEXT         NOT NULL,
+    price       REAL         NOT NULL
 );
